requirements.txt
click==7.1.2
kafka-python==2.0.2
psycopg2-binary==2.8.6
requests==2.25.1
